+++ cpu/kinetis_common/spi.c
@@ -464,12 +464,12 @@ static spi_state_t spi_config[SPI_NUMOF];
  * @return The actual achieved frequency on success
  * @return Less than 0 on error.
  */
-static int find_closest_baudrate_scalers(const uint32_t module_clock, const uint32_t target_clock,
+static long find_closest_baudrate_scalers(const uint32_t module_clock, const long target_clock,
         uint8_t *closest_prescaler, uint8_t *closest_scaler)
 {
     uint8_t i;
     uint8_t k;
-    int freq;
+    long freq;
     static const uint8_t num_scalers = 16;
     static const uint8_t num_prescalers = 4;
     static const int br_scalers[16] = {
@@ -478,7 +478,7 @@ static int find_closest_baudrate_scalers(const uint32_t module_clock, const uint
     };
     static const int br_prescalers[4] = {2, 3, 5, 7};
 
-    int closest_frequency = -1;
+    long closest_frequency = -1;
 
     /* Test all combinations until we arrive close to the target clock */
     for (i = 0; i < num_prescalers; ++i) {
@@ -533,18 +533,18 @@ static int find_closest_baudrate_scalers(const uint32_t module_clock, const uint
  * @return The actual achieved frequency on success
  * @return Less than 0 on error.
  */
-static int find_closest_delay_scalers(const uint32_t module_clock, const uint32_t target_freq,
+static long find_closest_delay_scalers(const uint32_t module_clock, const long target_freq,
                                       uint8_t *closest_prescaler, uint8_t *closest_scaler)
 {
     uint8_t i;
     uint8_t k;
-    int freq;
+    long freq;
     int prescaler;
     int scaler;
     static const uint8_t num_scalers = 16;
     static const uint8_t num_prescalers = 4;
 
-    int closest_frequency = -1;
+    long closest_frequency = -1;
 
     /* Test all combinations until we arrive close to the target clock */
     for (i = 0; i < num_prescalers; ++i) {
@@ -1041,13 +1041,13 @@ int spi_transfer_bytes(spi_t dev, char *out, char *in, unsigned int length)
     /* Default: send idle data */
     byte_out = (uint8_t)SPI_IDLE_DATA;
 
-    for (i = 0; i < length; i++) {
+    for (i = 0; i < (int)length; i++) {
         if (out != NULL) {
             /* Send given out data */
             byte_out = (uint8_t)out[i];
         }
 
-        if (i >= length - 1) {
+        if (i >= (int)length - 1) {
             /* Last byte, set End-of-Queue flag, clear Continue flag. */
             flags &= ~(SPI_PUSHR_CONT_MASK);
             flags |= SPI_PUSHR_EOQ_MASK;
@@ -1248,13 +1248,13 @@ int spi_transfer_regs(spi_t dev, uint8_t reg, char *out, char *in, unsigned int 
     /* Default: send idle data */
     byte_out = (uint8_t)SPI_IDLE_DATA;
 
-    for (i = 0; i < length; i++) {
+    for (i = 0; i < (int)length; i++) {
         if (out != NULL) {
             /* Send given out data */
             byte_out = (uint8_t)out[i];
         }
 
-        if (i >= length - 1) {
+        if (i >= (int)length - 1) {
             /* Last byte, set End-of-Queue flag, clear Continue flag. */
             flags &= ~(SPI_PUSHR_CONT_MASK);
             flags |= SPI_PUSHR_EOQ_MASK;
